Add GUI functionality for Events
Originally, in the base app, there was only 1 list that displays
Volunteer. This adds an extra column to the main window of the GUI such
that it can display events. This involves the addition of a few classes,
namely EventCard, EventListPanel.

Refactored some other volunteer GUI elements as well.

Changed the title of the GUI app to iVolunteer.

Also fixed a bug where events don't load from EventStorage due to the
DateTime format being different from LocalDateTime that is saved into
EventStorage. This is fixed through the DateTime::toString() method
to save the DateTime in the format as specified in the ecreate command.